Update init.lua
Greatly improved performance on MacBook by introducing a new mechanism:
Instead of scanning the Excel UI continuously, even while you are on Google or in a game, the scan will only start after you press the Option key on your keyboard. When the Option key is pressed, the program first checks if Excel is open. If it is, it will then scan the UI, but only for 5 seconds. More than enough time to open the different menus while also improving performance both inside Excel and in other applications.